installer: Provide event handlers with the event data.
* gnu/system/installer/configure.scm (configure-page-activate-selected-item):
Rename to...
(configure-page-activate-item): ...this.  And adapt it to use the passed
event data.
(make-configure-page): Use it here.
* gnu/system/installer/disks.scm (disk-page-activate-selected-item):
Rename to...
(disk-page-activate-item): ...this.  And adapt it to use the passed
event data.
(make-disk-page): Use it here.
* gnu/system/installer/filesystems.scm
(filesystem-page-activate-selected-item): Rename to...
(filesystem-page-activate-item): ...this.  And adapt it to use the passed
event data.
(make-filesystem-page): Use it here.
* gnu/system/installer/format.scm (format-page-activate-selected-item):
Rename to...
(format-page-activate-item): ...this.  And adapt it to use the passed
event data.
(make-format-page): Use it here.
* gnu/system/installer/guixsd-installer.scm
(main-page-activate-selected-item): Rename to...
(main-page-activate-item): ...this.  And adapt it to use the passed
event data.
(guixsd-installer): Use it here.
* gnu/system/installer/hostname.scm (host-name-mouse-handler):
Delete variable.
(host-name-key-handler): Delete variable.
(host-name-activate-item): New variable.  Lose hostname length limit check
for now.
(make-host-name-page): Use it here.
* gnu/system/installer/install.scm (install-page-mouse-handler): Delete
variable.
(install-page-key-handler): Delete variable.
(install-page-activate-item): New variable.
(make-install-page): Use it here.
* gnu/system/installer/key-map.scm (key-map-page-activate-selected-item):
Rename to...
(key-map-page-activate-item): ...this.  And adapt it to use the passed
event data.
(make-key-map): Use it here.
* gnu/system/installer/locale.scm (locale-page-activate-selected-item):
Rename to...
(locale-page-activate-item): ...this.  And adapt it to use the passed
event data.
(make-locale-page): Use it here.
* gnu/system/installer/mount-point.scm
(mount-point-page-activate-selected-item): Rename to...
(mount-point-page-activate-item): ...this.  And adapt it to use the passed
event data.  Export it.
* gnu/system/installer/network.scm (network-page-activate-selected-item):
Rename to...
(network-page-activate-item): ...this.  And adapt it to use the passed
event data.
(make-network-page): Use it here.
* gnu/system/installer/page.scm (page-activate-selected-item):
Rename to....
(page-activate-item): ...this.  And adapt it to use the passed event
data.
(page-default-mouse-handler): Simplify implementation.
(page-default-key-handler): Simplify implementation.
* gnu/system/installer/ping.scm (ping-page-activate-selected-item):
Rename to...
(ping-page-activate-item): ...this.  And adapt it to use the passed
event data.  Export it.
* gnu/system/installer/role.scm (role-page-activate-selected-item):
Rename to...
(role-page-activate-item): ...this.  And adapt it to use the passed
event data.
(make-role-page): Use it here.
* gnu/system/installer/time-zone.scm
(time-zone-page-activate-selected-item): Rename to...
(time-zone-page-activate-item): ...this.  And adapt it to use the passed
event data.
(make-tz-browser): Use it here.
* gnu/system/installer/user-edit.scm
(user-edit-page-activate-selected-item): Rename to...
(user-edit-page-activate-item): ...this.  And adapt it to use the passed
event data.
(make-user-edit-page): Use it here.
* gnu/system/installer/users.scm (users-page-activate-selected-item):
Rename to...
(users-page-activate-item): ...this.  And adapt it to use the passed
event data.
(make-users-page): Use it here.
* gnu/system/installer/wireless.scm (wireless-page-activate-selected-item):
Rename to...
(wireless-page-activate-item): ...this.  And adapt it to use the passed
event data.
(make-wireless-page): Use it here.
* gurses/menu.scm (std-menu-key-handler): Replace 'ignored by #f.
(std-menu-mouse-handler): Replace 'ignored by #f.